Extended Support for YUVU images
I needed to create a JPEG image from an YUVU image in memory, directly, without the overhead of transcoding in RGB back and forth. I Tryed to keep the modifications to your clean and nice code as little as possible... take a look and see if could be useful to others. Not extensively tested, at the moment.